+++ lib/domain_config.go
@@ -11,33 +11,46 @@ import (
 	feemarkettypes "github.com/skip-mev/feemarket/x/feemarket/types"
 )
 
+const AutoGasPrices = "auto"
+
+// Minimum values
 const (
 	WindowCorrectionFactorSuggestedMin = 0.5
 	BlockDurationEstimatedMin          = 1.0
 	GasPriceUpdateIntervalMin          = 5
 	RetryDelayMin                      = 1
 	AccountSequenceRetryDelayMin       = 1
-	AutoGasPrices                      = "auto"
+)
+
+// Default values
+const (
+	DefaultTimeoutRPCSecondsQuery        int64 = 60
+	DefaultTimeoutRPCSecondsTx           int64 = 300
+	DefaultTimeoutRPCSecondsRegistration int64 = 300
+	DefaultGasPriceUpdateInterval        int64 = 60
 )
 
 // Properties manually provided by the user as part of UserConfig
 type WalletConfig struct {
-	Address                   string // will be overwritten by the keystore. This is the 1 value that is auto-generated in this struct
-	AddressKeyName            string // load a address by key from the keystore
-	AddressRestoreMnemonic    string
-	AlloraHomeDir             string  // home directory for the allora keystore
-	Gas                       string  // gas to use for the allora client
-	GasAdjustment             float64 // gas adjustment to use for the allora client
-	GasPrices                 string  // gas prices to use for the allora client - "auto" for no fees
-	GasPriceUpdateInterval    int64   // number of seconds to wait between updates to the gas price
-	MaxFees                   uint64  // max gas to use for the allora client
-	NodeRpc                   string  // rpc node for allora chain
-	MaxRetries                int64   // retry to get data from chain up to this many times per query or tx
-	RetryDelay                int64   // number of seconds to wait between retries (general case)
-	AccountSequenceRetryDelay int64   // number of seconds to wait between retries in case of account sequence error
-	SubmitTx                  bool    // useful for dev/testing. set to false to run in dry-run processes without committing to the chain
-	BlockDurationEstimated    float64 // estimated average block duration in seconds
-	WindowCorrectionFactor    float64 // correction factor for the time estimation, suggested range 0.7-0.9.
+	Address                       string // will be overwritten by the keystore. This is the 1 value that is auto-generated in this struct
+	AddressKeyName                string // load a address by key from the keystore
+	AddressRestoreMnemonic        string
+	AlloraHomeDir                 string  // home directory for the allora keystore
+	Gas                           string  // gas to use for the allora client
+	GasAdjustment                 float64 // gas adjustment to use for the allora client
+	GasPrices                     string  // gas prices to use for the allora client - "auto" for auto-calculated fees
+	GasPriceUpdateInterval        int64   // number of seconds to wait between updates to the gas price
+	MaxFees                       uint64  // max fees to pay for a single transaction
+	NodeRpc                       string  // rpc node for allora chain
+	MaxRetries                    int64   // retry to get data from chain up to this many times per query or tx
+	RetryDelay                    int64   // number of seconds to wait between retries (general case)
+	AccountSequenceRetryDelay     int64   // number of seconds to wait between retries in case of account sequence error
+	SubmitTx                      bool    // useful for dev/testing. set to false to run in dry-run processes without committing to the chain
+	BlockDurationEstimated        float64 // estimated average block duration in seconds
+	WindowCorrectionFactor        float64 // correction factor for the time estimation, suggested range 0.7-0.9.
+	TimeoutRPCSecondsQuery        int64   // timeout for rpc queries in seconds, including retries
+	TimeoutRPCSecondsTx           int64   // timeout for rpc data send in seconds, including retries
+	TimeoutRPCSecondsRegistration int64   // timeout for rpc registration in seconds, including retries
 }
 
 // Properties auto-generated based on what the user has provided in WalletConfig fields of UserConfig
@@ -129,6 +142,22 @@ type ValueBundle struct {
 	OneOutInfererValues    []NodeValue `json:"oneOutInfererValues,omitempty"`
 	OneOutForecasterValues []NodeValue `json:"oneOutForecasterValues,omitempty"`
 	OneInForecasterValues  []NodeValue `json:"oneInForecasterValues,omitempty"`
+}
+
+// Check and set defaults for the user config if any values are not set
+func (c *UserConfig) CheckAndSetDefaults() {
+	if c.Wallet.TimeoutRPCSecondsQuery == 0 {
+		c.Wallet.TimeoutRPCSecondsQuery = DefaultTimeoutRPCSecondsQuery
+	}
+	if c.Wallet.TimeoutRPCSecondsTx == 0 {
+		c.Wallet.TimeoutRPCSecondsTx = DefaultTimeoutRPCSecondsTx
+	}
+	if c.Wallet.TimeoutRPCSecondsRegistration == 0 {
+		c.Wallet.TimeoutRPCSecondsRegistration = DefaultTimeoutRPCSecondsRegistration
+	}
+	if c.Wallet.GasPriceUpdateInterval == 0 {
+		c.Wallet.GasPriceUpdateInterval = DefaultGasPriceUpdateInterval
+	}
 }
 
 // Check that each assigned entrypoint in the user config actually can be used