+++ main.go
@@ -106,7 +106,7 @@ func main() {
 	// Convert entrypoints to instances of adapters
 	err := ConvertEntrypointsToInstances(finalUserConfig)
 	if err != nil {
-		log.Fatal().Err(err).Msg("Failed to convert Entrypoints to instances of adapters")
+		log.Fatal().Err(err).Msg("Failed to convert Entrypoints to instances of adapters - wrong entrypoint name?")
 		return
 	}
 	spawner, err := usecase.NewUseCaseSuite(finalUserConfig)